HybridJobSearchChatBot/requirement.txt
streamlit
langchain
groq
python-dotenv
PyMuPDF
faiss-cpu
transformers
sentence-transformers
